+++ src/main/java/duke/commands/Command.java
@@ -1,3 +1,8 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+
 abstract public class Command {
     private String desc;
     public Command (String desc) {
@@ -7,6 +12,6 @@ abstract public class Command {
     public String getDesc() {
         return this.desc;
     }
-    abstract boolean isExit();
-    abstract void execute(TaskList tasks, Storage storage);
+    public abstract boolean isExit();
+    public abstract void execute(TaskList tasks, Storage storage);
 }
