+++ src/main/java/duke/commands/UnknownCommand.java
@@ -1,15 +1,20 @@
+package duke.commands;
+
+import duke.Storage;
+import duke.TaskList;
+
 public class UnknownCommand extends Command {
     public UnknownCommand(String desc) {
         super(desc);
     }
 
     @Override
-    boolean isExit() {
+    public boolean isExit() {
         return false;
     }
 
     @Override
-    void execute(TaskList tasks, Storage storage) {
+    public void execute(TaskList tasks, Storage storage) {
         System.out.println("*** Apologies, Master Wayne. But I don't know what that means ***");
     }
 }
