+++ src/main/java/duke/tasks/Deadline.java
@@ -1,3 +1,5 @@
+package duke.tasks;
+
 public class Deadline extends Task {
 
     protected String by;
