+++ src/main/java/duke/tasks/Todo.java
@@ -1,3 +1,7 @@
+package duke.tasks;
+
+import duke.tasks.Task;
+
 public class Todo extends Task {
 
     public Todo(String description, boolean isDone) {
